[etc] 코드 스타일 수정
