+++ src/Constant/MESSAGE.js
@@ -1,0 +1,6 @@
+export const MESSAGE = {
+  START: "경주 할 자동차 이름(이름은 쉼표(,) 기준으로 구분)\n",
+  ATTEMPT: "시도할 횟수는 몇 회인가요?\n",
+  RESULT: "실행 결과",
+  END: "최종 우승자 : ",
+};
