+++ src/Constant/NUMBER.js
@@ -1,0 +1,5 @@
+export const NUMBER = {
+  MIN: 0,
+  MAX: 9,
+  STANDARD: 4,
+};
