+++ src/Game/getParticipant.js
@@ -1,0 +1,12 @@
+import { MESSAGE } from '../Constant/MESSAGE';
+import { readLineAsync } from '../Utils/readLineAsync';
+import { verifyParticipants } from '../verify';
+
+export const getParticipant = async () => {
+  const participants = await readLineAsync(MESSAGE.START);
+  const participantsList = participants.split(',');
+
+  await verifyParticipants(participantsList);
+
+  return participantsList;
+};
