+++ src/Utils/print.js
@@ -1,0 +1,6 @@
+import { MissionUtils } from '@woowacourse/mission-utils';
+const { Console } = MissionUtils;
+
+export const print = (message) => {
+  Console.print(message);
+};
