+++ src/Utils/readLineAsync.js
@@ -1,0 +1,6 @@
+import { MissionUtils } from '@woowacourse/mission-utils';
+const { Console } = MissionUtils;
+
+export const readLineAsync = async (message) => {
+  return await Console.readLineAsync(message);
+};
